Restructure code mostly for read only models
