+++ mso/base_model.py
@@ -16,26 +16,9 @@ from datetime import datetime
 from bson import ObjectId
 from decimal import Decimal
 from typing import Any, List, Dict
+from mso import config
 
-# ----------------------------------------------------------------------------------
-# BSON Type Mapping (used to validate types from MongoDB schema)
-# ----------------------------------------------------------------------------------
-BSON_TYPE_MAP = {
-    "string": str,
-    "int": int,
-    "bool": bool,
-    "double": float,
-    "float": float,
-    "date": datetime,
-    "objectId": ObjectId,
-    "binData": bytes,
-    "decimal": Decimal,
-    "long": int,
-    "timestamp": datetime,
-    "null": type(None),
-    "object": dict,
-    "array": list,
-}
+
 # ----------------------------------------------------------------------------------
 # ListFieldWrapper - A custom list wrapper for MongoDB array fields
 # ----------------------------------------------------------------------------------
@@ -302,9 +285,9 @@ class MongoModel:
         # MULTI-TYPE SUPPORT
         # ----------------------------
         if isinstance(bson_type, list):
-            expected_types = [BSON_TYPE_MAP.get(t, object) for t in bson_type if t in BSON_TYPE_MAP]
+            expected_types = [config.BSON_TYPE_MAP.get(t, object) for t in bson_type if t in config.BSON_TYPE_MAP]
         else:
-            expected_types = [BSON_TYPE_MAP.get(bson_type, object)]
+            expected_types = [config.BSON_TYPE_MAP.get(bson_type, object)]
 
         # ----------------------------
         # OBJECT VALIDATION
@@ -339,7 +322,7 @@ class MongoModel:
                     continue
 
                 # SCALAR VALIDATION
-                expected_item_type = BSON_TYPE_MAP.get(item_type, object)
+                expected_item_type = config.BSON_TYPE_MAP.get(item_type, object)
                 if not isinstance(item, expected_item_type):
                     # Attempt coercion since initial type check failed
                     value = self.coerce_type(item, [expected_item_type])
