+++ mso/config.py
@@ -11,43 +11,27 @@
 #                                                                                                                      #
 #  Gitlab: https://github.com/chuckbeyor101/MSO-Mongo-Schema-Object-Library                                            #
 # ######################################################################################################################
-from typing import Any
-from mso import base_model, config
-from pymongo.database import Database
 
+from datetime import datetime
+from decimal import Decimal
+from bson import ObjectId
 
-def parse_schema(schema, name='Root'):
-    classes = {}
-    properties = schema.get('properties', {})
-    for prop_name, prop_info in properties.items():
-        if prop_info.get('type') == 'object':
-            nested_class_name = f"{name}_{prop_name.capitalize()}"
-            nested_classes = parse_schema(prop_info, nested_class_name)
-            classes.update(nested_classes)
-    classes[name] = schema
-    return classes
-
-
-def normalize_class_name(name):
-    return name.replace(' ', '_').replace('-', '_')
-
-
-def normalize_bson_type(bson_type):
-    if isinstance(bson_type, list):
-        for t in bson_type:
-            if t != 'null':
-                return t
-    return bson_type
-
-
-def infer_python_type_from_bson(bson_type):
-    if isinstance(bson_type, list):
-        for t in bson_type:
-            if t != "null":
-                return config.BSON_TYPE_MAP.get(t, Any)
-        return Any
-    return config.BSON_TYPE_MAP.get(bson_type, Any)
-
-
-def is_view(db: Database, collection_name: str) -> bool:
-    return db["system.views"].find_one({"_id": f"{db.name}.{collection_name}"}) is not None
+# ----------------------------------------------------------------------------------
+# BSON Type Mapping (used to validate types from MongoDB schema)
+# ----------------------------------------------------------------------------------
+BSON_TYPE_MAP = {
+    "string": str,
+    "int": int,
+    "bool": bool,
+    "double": float,
+    "float": float,
+    "date": datetime,
+    "objectId": ObjectId,
+    "binData": bytes,
+    "decimal": Decimal,
+    "long": int,
+    "timestamp": datetime,
+    "null": type(None),
+    "object": dict,
+    "array": list,
+}
