+++ pyproject.toml
@@ -32,6 +32,7 @@ pymongo = "^4.0"
 fastapi = "^0.110.0"
 uvicorn = { extras = ["standard"], version = "^0.27.0" }
 pydantic = "^2.6.0"
+python-dateutil = "^2.9.0"
 
 [project.urls]
 Pypi = "https://pypi.org/project/MSO/"
